requirements.txt
wave
webrtcvad
array
pydub
subprocess
shlex
statistics
librosa
matplotlib.pyplot
numpy
soundfile
pandas
pyloudnorm
moviepy